+++ src/lb.cpp
@@ -58,7 +58,7 @@ void zmq::lb_t::terminate ()
     zmq_assert (!terminating);
     terminating = true;
 
-    sink->register_term_acks (pipes.size ());
+    sink->register_term_acks ((int) pipes.size ());
     for (pipes_t::size_type i = 0; i != pipes.size (); i++)
         pipes [i]->terminate ();
 }
